Refactor the intervalindex-related API (#16573)
[AB#5050](https://dev.azure.com/fluidframework/internal/_workitems/edit/5050)

1. Move the interfaces and implementations of different IntervalIndex to
`./intervalIndex`, make the `intervalCollection.ts` solely focuses on
the `IntervalCollection` itself
2. Deprecate some APIs related to the intervalIndex, We would like the
user to attach the index to the collection on their own.

---------
