+++ packages/dds/sequence/src/index.ts
@@ -32,21 +32,25 @@ export {
 export {
 	DeserializeCallback,
 	IIntervalCollectionEvent,
-	IntervalIndex,
 	IIntervalCollection,
 	IntervalLocator,
 	intervalLocatorFromEndpoint,
-	IEndpointInRangeIndex,
-	IStartpointInRangeIndex,
-	createEndpointInRangeIndex,
-	createStartpointInRangeIndex,
 } from "./intervalCollection";
 export { IntervalConflictResolver } from "./intervalTree";
 export {
+	IntervalIndex,
 	SequenceIntervalIndexes,
 	IOverlappingIntervalsIndex,
 	createOverlappingIntervalsIndex,
 	createOverlappingSequenceIntervalsIndex,
+	IEndpointInRangeIndex,
+	IStartpointInRangeIndex,
+	createEndpointInRangeIndex,
+	createStartpointInRangeIndex,
+	IIdIntervalIndex,
+	createIdIntervalIndex,
+	IEndpointIndex,
+	createEndpointIndex,
 } from "./intervalIndex";
 export {
 	appendAddIntervalToRevertibles,
