+++ packages/dds/sequence/src/intervalIndex/index.ts
@@ -3,6 +3,11 @@
  * Licensed under the MIT License.
  */
 
+export { IntervalIndex } from "./intervalIndex";
+export { IIdIntervalIndex, createIdIntervalIndex } from "./idIntervalIndex";
+export { IEndpointIndex, createEndpointIndex } from "./endpointIndex";
+export { IEndpointInRangeIndex, createEndpointInRangeIndex } from "./endpointInRangeIndex";
+export { IStartpointInRangeIndex, createStartpointInRangeIndex } from "./startpointInRangeIndex";
 export { SequenceIntervalIndexes } from "./sequenceIntervalIndexes";
 export {
 	IOverlappingIntervalsIndex,
