+++ packages/dds/sequence/src/test/endpointInRangeIndex.spec.ts
@@ -6,7 +6,7 @@
 import { strict as assert } from "assert";
 import { makeRandom } from "@fluid-internal/stochastic-test-utils";
 import { Client } from "@fluidframework/merge-tree";
-import { IEndpointInRangeIndex, createEndpointInRangeIndex } from "../intervalCollection";
+import { IEndpointInRangeIndex, createEndpointInRangeIndex } from "../intervalIndex";
 import { Interval, intervalHelpers } from "../intervals";
 import {
 	assertPlainNumberIntervals,
